requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# Web Framework
flask==3.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# AI Services
elevenlabs==0.2.27
openai==1.6.1

# Image Processing
pillow==10.1.0

# Database
sqlalchemy==2.0.23

# Payment Processing
stripe==7.8.0

# Environment Variables
python-dotenv==1.0.0

# Task Scheduling
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
gunicorn==21.2.0

# Promo/Marketing
praw>=7.7.0
